handle pools & remove dead code for treasury
