+++ src/api/treasury/getTreasury.ts
@@ -1,29 +1,23 @@
 import BigNumber from 'bignumber.js';
-import { keyBy } from 'lodash';
 import { addressBook } from '../../../packages/address-book/src/address-book';
 import chainIdMap from '../../../packages/address-book/src/util/chainIdMap';
 import { getKey, setKey } from '../../utils/cache';
-import { getSingleChainVaults } from '../stats/getMultichainVaults';
 import { extractBalancesFromTreasuryCallResults, mapAssetToCall } from './multicallUtils';
+import { getTokenAddressesByChain, getVaultAddressesByChain } from './assetHelpers';
 import {
+  isGovAsset,
   isValidatorAsset,
   isVaultAsset,
-  MarketMakerAPIResult,
-  MMExchangeBalance,
-  MMReport,
   TreasuryAsset,
   TreasuryAssetRegistry,
   TreasuryBalances,
   TreasuryReport,
   TreasuryWalletRegistry,
 } from './types';
-import { getChainValidators, hasChainValidator } from './validatorHelpers';
 import { serviceEventBus } from '../../utils/ServiceEventBus';
 import { ApiChain, SupportedChains } from '../../utils/chain';
-import { getTokensForChain, isTokenNative } from '../tokens/tokens';
 import { getAmmPrice } from '../stats/getAmmPrices';
 import { keysToObject } from '../../utils/array';
-import { getChainConcentratedLiquidityAssets, hasChainConcentratedLiquidityAssets } from './nftAssets';
 import { ZERO_ADDRESS } from '../../utils/address';
 import { contextAllSettled, isContextResultRejected, withTimeout } from '../../utils/promise';
 import { envNumber } from '../../utils/env';
@@ -39,9 +33,6 @@ let assetsByChain: TreasuryAssetRegistry = keysToObject(SupportedChains, () => (
 let tokenBalancesByChain: TreasuryBalances = keysToObject(SupportedChains, () => ({}));
 
 let treasurySummary: TreasuryReport = keysToObject(SupportedChains, () => ({}));
-
-// market maker
-let marketMakerReport: MMReport = {};
 
 function updateTreasuryAddressesByChain() {
   treasuryAddressesByChain = keysToObject(SupportedChains, chain => {
@@ -78,74 +69,6 @@ function updateAssetsByChain() {
       ...(tokenAssets[chain] || {}),
       ...(vaultAssets[chain] || {}),
     };
-  });
-}
-
-// Load token address
-function getTokenAddressesByChain(): TreasuryAssetRegistry {
-  return keysToObject(SupportedChains, chain => {
-    const tokens: Record<string, TreasuryAsset> = {};
-
-    for (const [tokenAddress, token] of Object.entries(getTokensForChain(chain))) {
-      // WCELO/WMETIS/WGLMR: duplicate as same as native
-      if (
-        [
-          '0x471EcE3750Da237f93B8E339c536989b8978a438',
-          '0xDeadDeAddeAddEAddeadDEaDDEAdDeaDDeAD0000',
-          '0x0000000000000000000000000000000000000802',
-        ].includes(token.address)
-      )
-        continue;
-
-      tokens[tokenAddress] = {
-        name: token.name,
-        address: token.address,
-        decimals: token.decimals,
-        assetType: isTokenNative(token) ? 'native' : 'token',
-        oracleId: token.oracleId,
-        oracleType: token.oracle,
-        symbol: token.symbol,
-        ...(token.staked && { staked: true }),
-      };
-    }
-
-    if (hasChainValidator(chain)) {
-      const validators = getChainValidators(chain);
-      for (let i = 0; i < validators.length; i++) {
-        tokens[validators[i].id] = validators[i];
-      }
-    }
-
-    if (hasChainConcentratedLiquidityAssets(chain)) {
-      getChainConcentratedLiquidityAssets(chain).forEach(asset => {
-        tokens[asset.address.toLowerCase()] = {
-          ...asset,
-          staked: true,
-        };
-      });
-    }
-
-    return tokens;
-  });
-}
-
-function getVaultAddressesByChain(): TreasuryAssetRegistry {
-  return keysToObject(SupportedChains, chain => {
-    const chainVaults = getSingleChainVaults(chain) || [];
-    return keyBy(
-      chainVaults.map(vault => ({
-        name: vault.earnedToken,
-        oracleId: vault.oracleId,
-        oracleType: vault.oracle,
-        assetType: 'vault',
-        vaultId: vault.id,
-        decimals: vault.tokenDecimals,
-        pricePerFullShare: vault.pricePerFullShare,
-        address: vault.earnContractAddress,
-        staked: true,
-      })),
-      asset => asset.address.toLowerCase()
-    );
   });
 }
 
@@ -278,7 +201,7 @@ async function buildTreasuryReportForChain(chain: ApiChain): Promise<TreasuryRep
 }
 
 function findUsdValueForBalance(tokenInfo: TreasuryAsset, tokenPrice: number, balance: BigNumber): BigNumber {
-  if (isVaultAsset(tokenInfo)) {
+  if (isVaultAsset(tokenInfo) || isGovAsset(tokenInfo)) {
     return balance
       .multipliedBy(tokenPrice)
       .multipliedBy(tokenInfo.pricePerFullShare)
@@ -292,13 +215,11 @@ function findUsdValueForBalance(tokenInfo: TreasuryAsset, tokenPrice: number, ba
 async function saveToRedis() {
   await setKey('TREASURY_BALANCES', tokenBalancesByChain);
   await setKey('TREASURY_REPORT', treasurySummary);
-  await setKey('MM_REPORT', marketMakerReport);
 }
 
 async function restoreFromRedis() {
   const cachedSummary = await getKey<TreasuryReport>('TREASURY_REPORT');
   const cachedBalances = await getKey<TreasuryBalances>('TREASURY_BALANCES');
-  const cachedMMReport = await getKey<MMReport>('MM_REPORT');
   if (cachedSummary) {
     treasurySummary = cachedSummary;
   }
@@ -312,9 +233,6 @@ async function restoreFromRedis() {
         });
       });
     });
-  }
-  if (cachedMMReport) {
-    marketMakerReport = cachedMMReport;
   }
 }
 
@@ -343,15 +261,4 @@ export async function initTreasuryService() {
 
 export function getBeefyTreasury(): TreasuryReport {
   return treasurySummary;
-}
-
-export function getMarketMakerBalances(): MMReport {
-  return marketMakerReport;
-}
-
-export function getAllBeefyHoldings() {
-  return {
-    treasury: getBeefyTreasury(),
-    marketMaker: getMarketMakerBalances(),
-  };
 }
