+++ src/api/treasury/multicallUtils.ts
@@ -3,6 +3,7 @@ import {
   AssetBalance,
   ChainTreasuryBalance,
   isConcLiquidityAsset,
+  isGovAsset,
   isNativeAsset,
   isTokenAsset,
   isValidatorAsset,
@@ -18,12 +19,8 @@ import { MULTICALL_V3 } from '../../utils/multicallHelpers';
 import MulticallAbi from '../../abis/common/Multicall/MulticallAbi';
 import { getValidatorBalanceCall } from './validatorHelpers';
 
-export const mapAssetToCall = (
-  asset: TreasuryAsset,
-  treasuryAddressesForChain: TreasuryWallet[],
-  chainId: number
-) => {
-  if (isTokenAsset(asset) || isVaultAsset(asset)) {
+export const mapAssetToCall = (asset: TreasuryAsset, treasuryAddressesForChain: TreasuryWallet[], chainId: number) => {
+  if (isTokenAsset(asset) || isVaultAsset(asset) || isGovAsset(asset)) {
     const contract = fetchContract(asset.address, ERC20Abi, chainId);
     return treasuryAddressesForChain.map(treasuryData =>
       contract.read.balanceOf([treasuryData.address as `0x${string}`])
@@ -48,10 +45,8 @@ export const extractBalancesFromTreasuryCallResults = (
   const allBalances: AssetBalance[] = [];
   apiAssets.forEach((asset, i) => {
     if (callResults[i].status === 'fulfilled') {
-      const callResult = callResults[i] as PromiseFulfilledResult<
-        bigint[] | LpBreakdown[] | TreasuryApiResult[]
-      >;
-      if (isTokenAsset(asset) || isVaultAsset(asset) || isNativeAsset(asset)) {
+      const callResult = callResults[i] as PromiseFulfilledResult<bigint[] | LpBreakdown[] | TreasuryApiResult[]>;
+      if (isTokenAsset(asset) || isVaultAsset(asset) || isGovAsset(asset) || isNativeAsset(asset)) {
         const value = callResult.value as bigint[];
         const bal = {
           address: asset.address.toLowerCase(),
